Move Spline classes into geom.py
Why on the Earth do we have four implementations of cubic_spline()
method? There was also fifth implementation by ly29, more OOP-like.
