+++ utils/geom.py
@@ -434,30 +434,74 @@ rects = vectorize(rect)
 lines = vectorize(line)
 grids = vectorize(grid)
 
+################################################
+# Newer implementation of spline interpolation
+# by zeffii, ly29 and portnov
+# based on implementation from looptools 4.5.2 done by Bart Crouch
+# factored out from interpolation_mk3 node
+################################################
 
-# ---------- Spline
+class Spline(object):
+    """
+    Base abstract class for LinearSpline and CubicSpline.
+    """
+    @classmethod
+    def create_knots(cls, pts, metric="DISTANCE"):
+        if metric == "DISTANCE":
+            tmp = np.linalg.norm(pts[:-1] - pts[1:], axis=1)
+            tknots = np.insert(tmp, 0, 0).cumsum()
+            tknots = tknots / tknots[-1]
+        elif metric == "MANHATTAN":
+            tmp = np.sum(np.absolute(pts[:-1] - pts[1:]), 1)
+            tknots = np.insert(tmp, 0, 0).cumsum()
+            tknots = tknots / tknots[-1]
+        elif metric == "POINTS":
+            tknots = np.linspace(0, 1, len(pts))
+        elif metric == "CHEBYSHEV":
+            tknots = np.max(np.absolute(pts[1:] - pts[:-1]), 1)
+            tmp = np.insert(tmp, 0, 0).cumsum()
+            tknots = tknots / tknots[-1]
 
-# spline function modifed from
-# from looptools 4.5.2 done by Bart Crouch
+        return tknots
 
-
-# calculates natural cubic splines through all given knots
-
-class CubicSpline:
-    def __init__(self, locs, tknots=None, metric='DISTANCE'):
-        """    locs is and np.array with shape (n,3) and tknots has shape (n-1,)
-        creates a cubic spline thorugh the locations given in locs
-
+class CubicSpline(Spline):
+    def __init__(self, vertices, tknots = None, metric = None, is_cyclic = False):
         """
-        n = len(locs)
-        if n < 2:
-            return False
+        vertices: vertices in Sverchok's format (list of tuples)
+        tknots: np.array of shape (n-1,). If not provided - calculated automatically based on metric
+        metric: string, one of "DISTANCE", "MANHATTAN", "POINTS", "CHEBYSHEV". Mandatory if tknots
+                is not provided
+        is_cyclic: whether the spline is cyclic
 
-        if tknots is None:
-            tknots = create_knots(locs, metric)
+        creates a cubic spline thorugh the locations given in vertices
+        """
+
+        if is_cyclic:
+
+            locs = np.array(vertices[-4:] + vertices + vertices[:4])
+            if tknots is None:
+                if metric is None:
+                    raise Exception("CubicSpline: either tknots or metric must be specified")
+                tknots = Spline.create_knots(locs, metric)
+                scale = 1 / (tknots[-4] - tknots[4])
+                base = tknots[4]
+                tknots -= base
+                tknots *= scale
+        else:
+            locs = np.array(vertices)
+            if tknots is None:
+                if metric is None:
+                    raise Exception("CubicSpline: either tknots or metric must be specified")
+                tknots = Spline.create_knots(locs, metric)
 
         self.tknots = tknots
+        self.is_cyclic = is_cyclic
 
+        n = len(locs)
+        if n < 2:
+            return
+
+        # a = locs
         h = tknots[1:] - tknots[:-1]
         h[h == 0] = 1e-8
         q = np.zeros((n - 1, 3))
@@ -491,31 +535,35 @@ class CubicSpline:
         splines[:, 2] = c[:-1]
         splines[:, 3] = d
         splines[:, 4] = tknots[:-1, np.newaxis]
-
+        
         self.splines = splines
 
-
-    def eval(self, t_in):
+    def eval(self, t_in, tknots = None):
         """
         Evaluate the spline at the points in t_in, which must be an array
         with values in [0,1]
         returns and np array with the corresponding points
         """
-        splines = self.splines
-        tknots = self.tknots
+
+        if tknots is None:
+            tknots = self.tknots
+
         index = tknots.searchsorted(t_in, side='left') - 1
-        index = index.clip(0, len(splines) - 1)
-        to_calc = splines[index]
+        index = index.clip(0, len(self.splines) - 1)
+        to_calc = self.splines[index]
         ax, bx, cx, dx, tx = np.swapaxes(to_calc, 0, 1)
         t_r = t_in[:, np.newaxis] - tx
         out = ax + t_r * (bx + t_r * (cx + t_r * dx))
         return out
 
-
-    def tangent(self, t_in, h=0.001):
+    def tangent(self, t_in, h=0.001, tknots=None):
         """
         Calc numerical tangents for spline at t_in
         """
+
+        if tknots is None:
+            tknots = self.tknots
+
         t_ph = t_in + h
         t_mh = t_in - h
         t_less_than_0 = t_mh < 0.0
@@ -528,46 +576,103 @@ class CubicSpline:
         tanget[t_less_than_0 | t_great_than_1] *= 2
         return tanget
 
-def create_knots(pts, metric="DISTANCE"):
-    if metric == "DISTANCE":
-        tmp = np.linalg.norm(pts[:-1] - pts[1:], axis=1)
-        tknots = np.insert(tmp, 0, 0).cumsum()
-        tknots = tknots / tknots[-1]
-    elif metric == "MANHATTAN":
-        tmp = np.sum(np.absolute(pts[:-1] - pts[1:]), 1)
-        tknots = np.insert(tmp, 0, 0).cumsum()
-        tknots = tknots / tknots[-1]
-    elif metric == "POINTS":
-        tknots = np.linspace(0, 1, len(pts))
-    elif metric == "CHEBYSHEV":
-        tknots = np.max(np.absolute(pts[1:] - pts[:-1]), 1)
-        tmp = np.insert(tmp, 0, 0).cumsum()
-        tknots = tknots / tknots[-1]
+class LinearSpline(Spline):
+    def __init__(self, vertices, tknots = None, metric = None, is_cyclic = False):
+        """
+        vertices: vertices in Sverchok's format (list of tuples)
+        tknots: np.array of shape (n-1,). If not provided - calculated automatically based on metric
+        metric: string, one of "DISTANCE", "MANHATTAN", "POINTS", "CHEBYSHEV". Mandatory if tknots
+                is not provided
+        is_cyclic: whether the spline is cyclic
 
-    return tknots
+        creates a cubic spline thorugh the locations given in vertices
+        """
 
+        if is_cyclic:
+            pts = np.array(vertices + [vertices[0]])
+        else:
+            pts = np.array(vertices)
 
-class LinearSpline:
-    def __init__(self, pts, tknots=None, metric='DISTANCE'):
-        self.pts = np.array(pts).T
         if tknots is None:
-            tknots = create_knots(locs, metric)
+            if metric is None:
+                raise Exception("LinearSpline: either tknots or metric must be specified")
+            tknots = Spline.create_knots(pts, metric)
 
+        self.pts = pts
         self.tknots = tknots
+        self.is_cyclic = is_cyclic
 
-
-    def eval(self, t_in):
+    def eval(self, t_in, tknots = None):
         """
         Eval the liner spline f(t) = x,y,z through the points
         in pts given the knots in tknots at the point in t_in
         """
-        ptsT = self.pts
-        tknots = self.tknots
-        out = np.empty((3, len(t_in)))
+
+        if tknots is None:
+            tknots = self.tknots
+            
+        ptsT = self.pts.T
+        out = np.zeros((3, len(t_in)))
         for i in range(3):
             out[i] = np.interp(t_in, tknots, ptsT[i])
         return out.T
 
+    def tangent(self, t_in, tknots = None):
+        if tknots is None:
+            tknots = self.tknots
+
+        lookup_segments = GenerateLookup(self.is_cyclic, self.pts.tolist())
+        return [lookup_segments.find_bucket(f) for f in t_in]
+
+class GenerateLookup():
+
+    def __init__(self, cyclic, vlist):
+        self.lookup = {}
+        self.summed_lengths = []
+        self.indiv_lengths = []
+        self.normals = []
+        self.buckets = []
+        if cyclic:
+            vlist = vlist + [vlist[0]]
+
+        self.get_seq_len(vlist)
+        self.acquire_lookup_table()
+        self.get_buckets()
+        # for idx, (k, v) in enumerate(sorted(self.lookup.items())):
+        #     print(k, v)
+
+    def find_bucket(self, factor):
+        for bucket_min, bucket_max in zip(self.buckets[:-1], self.buckets[1:]):
+            if bucket_min <= factor < bucket_max:
+                tval = self.lookup.get(bucket_min)  # , self.lookup.get(self.buckets[-1]))
+
+                return tval
+
+        # return last bucket just in case
+        return self.lookup.get(self.buckets[-1])
+
+    def get_buckets(self):
+        self.buckets = [(clen / self.total_length) for clen in self.summed_lengths]
+    
+    def acquire_lookup_table(self):
+        for current_length, segment_normal in zip(self.summed_lengths, self.normals):
+            self.lookup[current_length / self.total_length] = segment_normal
+        
+    def get_seq_len(self, vlist):
+        add_len = self.indiv_lengths.append
+        add_normal = self.normals.append
+        add_to_sumlist = self.summed_lengths.append
+        current_length = 0.0
+        for idx in range(len(vlist)-1):
+            v = vlist[idx][0]-vlist[idx+1][0], vlist[idx][1]-vlist[idx+1][1], vlist[idx][2]-vlist[idx+1][2]
+            length = math.sqrt((v[0]*v[0]) + (v[1]*v[1]) + (v[2]*v[2]))
+            add_normal(v)
+            add_len(length)
+            add_to_sumlist(current_length)
+            current_length += length
+
+        self.total_length = sum(self.indiv_lengths)
+            
 
 def multiply_vectors(M, vlist):
     # (4*4 matrix)  X   (3*1 vector)
